extract <CompetitorsEditor>; share between settings + wizard
Both surfaces had their own implementation of editing competitors with
divergent UI:

- settings/competitors: collapsible card with pencil/trash icons,
  labels + tooltips per field, real domain validation
- prompt-wizard: bespoke flat card with one Input + read-only summary
  of additional domains/aliases (uneditable in the wizard)

Lift to a single controlled component (apps/web/src/components/
competitors-editor.tsx) with a `CompetitorEntry` type, an `onChange`
callback, and a `disabled` prop. Both pages render exactly the same
UI now; future tweaks happen in one place.

Side cleanup: settings was using a `[""]` sentinel for an empty domains
list; the new component takes a real `string[]` and keeps the field
editable via TagsInput's own empty state.
